requirements.txt
fastapi
uvicorn[standard]
httpx
requests
langchain
langchain-nvidia-ai-endpoints
langgraph
azure-cosmos
python-dotenv
pydantic
gunicorn
aiohttp
langchain-community
nvidia-riva-client
tavily-python
langchain-core
langchain-classic
langchain-text-splitters
colorama
# Google APIs
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
# Audio Processing
pydub==0.25.1
# Browser Automation
playwright==1.40.0
openai==1.12.0
# Google Gemini
google-generativeai==0.8.3
pillow==10.2.0
# Document Processing
unstructured[all-docs]
pytesseract
python-magic-bin
pdf2image
python-docx
python-pptx
openpyxl
pandas
aiofiles